+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -5,12 +5,15 @@ import lombok.Data;
 
 import javax.validation.constraints.*;
 import java.time.LocalDate;
+import java.util.Set;
 
 @Data
 @Builder
 public class User {
 
     private int id;
+
+    private Set<Long> friendVault;
 
     @Email
     @NotBlank
@@ -25,4 +28,5 @@ public class User {
     @PastOrPresent
     @NotNull
     private LocalDate birthday;
+
 }
